`useSuspenseQuery` modern and classic separated
